+++ comp2013/src/main/java/comp2013/SnakeMain.java
@@ -7,11 +7,11 @@ import javax.swing.text.View;
 
 public class SnakeMain extends Application {
     static SnakeModel M_Model;
-
-
+     static SnakeController M_Controller;
     // Initialise the Model, View and Controller.
     private static void initialise() {
         M_Model = new SnakeModel(700, 500);
+        M_Controller = new SnakeController();
     }
     public static void main(String[] args) {
         // Initialise the game.