+++ comp2013/src/main/java/comp2013/SnakeThread.java
@@ -3,6 +3,11 @@ package comp2013;
 import java.awt.*;
 import java.awt.Component;
 public class SnakeThread extends Thread{
+
+    private SnakeController M_Controller;
+    public SnakeThread(SnakeController controller){
+        this.M_Controller = controller;
+    }
     @Override
     public void run()
     {
@@ -10,6 +15,7 @@ public class SnakeThread extends Thread{
         while (true)
         {
             // Refresh the screen.
+            System.out.printf("Thread is running\n");
             try
             {
                 sleep(30);